# perf: optimize `MarshalFrame`